created handlebars and updated server to include required calls; moved api calls into folder and added routes to control homepage calls; adjust all necessary code to point to api folder and proper files
