+++ controller/api/taskRoutes.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const Task = require('../model/taskModel');
+const Task = require('../../model/taskModel'); //! onMergeCheck
 
 // Get all tasks
 router.get('/', async (req, res) => {
