+++ package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "express-handlebars": "^7.0.7",
     "mysql2": "^3.5.1",
     "sequelize": "^6.32.1",
     "uuid": "^9.0.0"
