+++ server.js
@@ -1,22 +1,22 @@
 const path = require('path');
 const express = require('express');
-//const exphbs = require('express-handlebars');
+const exphbs = require('express-handlebars');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
 
 const sequelize = require('./config/database');
 
-//const hbs = exphbs.create({});
+const hbs = exphbs.create({});
 
-//app.engine('handlebars', hbs.engine);
-//app.set('view engine', 'handlebars');
+app.engine('handlebars', hbs.engine);
+app.set('view engine', 'handlebars');
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use(require('./controller/'));
+app.use(require('./controller'));
 
 sequelize.sync({ force: false }).then(() => {
   app.listen(PORT, () => console.log('Now listening'));
