+++ views/layouts/main.handlebars
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Task</title> <!--? CAN we use {{variable}} for current task containing these "subtasks" view; nothing for root-->
+    <!--<link rel="stylesheet" type="text/css" href="???">-->
+</head>
+
+<body>
+    <header>
+    </header>
+    {{{body}}}
+</body>
+</html>
