+++ views/login.handlebars
@@ -1,0 +1,6 @@
+<!--[textbox] enter user name-->
+
+<!--[textbox] enter password-->
+
+<!--[button] sign up-->
+<!--[button] login-->
