+++ views/test.handlebars
@@ -1,0 +1,1 @@
+<h2> TEST </h2>
